+++ lib/qc/modules/data_wipe/providers/data_wipe_detail_provider.dart
@@ -7,11 +7,18 @@ import 'package:flutter_trc/qc/modules/data_wipe/resources/data_wipe_list_respon
 import 'package:flutter_trc/qc/modules/data_wipe/resources/data_wipe_service.dart';
 import 'package:provider/provider.dart';
 
+enum BottomButtonState { validation, scanAnother, initDataWipe }
+
 class DataWipeDetailProvider extends CshChangeNotifier {
   final String _deviceBarcode;
   DataWipeListItem? data;
   bool isLoading = true;
   bool forceHideInitiateButton = false;
+  final int eraserInfoFailedStatus = -11;
+  bool _isImeiSerialAlreadyUpdated = false;
+  BottomButtonState? _bottomButtonState;
+
+  bool get isImeiSerialAlreadyUpdated => _isImeiSerialAlreadyUpdated;
 
   DataWipeDetailProvider(this._deviceBarcode);
 
@@ -19,9 +26,25 @@ class DataWipeDetailProvider extends CshChangeNotifier {
     return Provider.of<DataWipeDetailProvider>(context, listen: listen);
   }
 
-  getDeviceWipeStatus({Function(String errorMessage)? onError}) {
+  BottomButtonState get bottomButtonState => _bottomButtonState ?? BottomButtonState.initDataWipe;
+
+  set bottomButtonState(BottomButtonState value) {
+    _bottomButtonState = value;
+    notifyListeners();
+  }
+
+  getDeviceWipeStatus({Function(String errorMessage)? onError, bool isFirstTime = false}) {
     DataWipeService.getDataWipeDetails(_deviceBarcode).listen((event) {
       data = event;
+      if (isFirstTime) {
+        if (data?.statusCode == eraserInfoFailedStatus) {
+          _bottomButtonState = BottomButtonState.validation;
+        } else if ((data?.statusCode ?? 0) < 1) {
+          _bottomButtonState = BottomButtonState.initDataWipe;
+        } else {
+          _bottomButtonState = BottomButtonState.scanAnother;
+        }
+      }
     }, onError: (error) {
       String? errorMessage = ApiErrorHelper.getErrorMessage(error);
       onError?.call(errorMessage.toString());
@@ -37,6 +60,20 @@ class DataWipeDetailProvider extends CshChangeNotifier {
       getDeviceWipeStatus();
       forceHideInitiateButton = true;
       completer.complete(true);
+    }, onError: (error) {
+      String? errorMessage = ApiErrorHelper.getErrorMessage(error);
+      completer.completeError(errorMessage.toString());
+    });
+    return completer.future;
+  }
+
+  Future<void> reportMisMatch({String? imei1, String? imei2, String? serialNo}) {
+    var completer = Completer<void>();
+    DataWipeService.reportMisMatch(_deviceBarcode, imei1: imei1, imei2: imei2, serialNo: serialNo).listen((_) {
+      _isImeiSerialAlreadyUpdated = true;
+      _bottomButtonState = BottomButtonState.initDataWipe;
+      completer.complete();
+      notifyListeners();
     }, onError: (error) {
       String? errorMessage = ApiErrorHelper.getErrorMessage(error);
       completer.completeError(errorMessage.toString());