+++ lib/qc/modules/data_wipe/resources/data_wipe_service.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 
+import 'package:core_widgets/core_widgets.dart';
 import 'package:flutter_trc/qc/modules/data_wipe/resources/data_wipe_filter_list_response.dart';
 import 'package:flutter_trc/qc/modules/data_wipe/resources/data_wipe_list_response.dart';
 import 'package:flutter_trc/src/common/model/base_action_response.dart';
@@ -30,5 +31,20 @@ class DataWipeService {
   static Stream<BaseActionResponse> bulkInitiate(int statusCode) {
     Map<String, dynamic> req = {"sc": statusCode};
     return QcService().post("/erasure-request/bulk-process", BaseActionResponse.fromJson, body: jsonEncode(req));
+  }
+
+  static Stream<BaseActionResponse> reportMisMatch(
+    String deviceBarcode, {
+    String? imei1,
+    String? imei2,
+    String? serialNo,
+  }) {
+    Map<String, dynamic> req = {
+      if (!Validator.isNullOrEmpty(imei1)) "imei": imei1,
+      if (!Validator.isNullOrEmpty(imei2)) "imei2": imei2,
+      if (!Validator.isNullOrEmpty(serialNo)) "sno": serialNo,
+    };
+    return QcService()
+        .post("/erasure-request/update/$deviceBarcode", BaseActionResponse.fromJson, body: jsonEncode(req));
   }
 }