+++ lib/src/common/widgets/imei_scanner.dart
@@ -1,20 +1,18 @@
 import 'dart:async';
 
 import 'package:builder_component/builder_component.dart';
-import 'package:camera/camera.dart';
-import 'package:core/core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_trc/src/app_builder/app_headers/qc_general_header/widgets/qc_general_header.dart';
 import 'package:flutter_trc/src/libraries/firebase/remote_config_helper.dart';
 import 'package:imei_serial_reader/imei_serial_reader.dart';
-import 'package:imei_serial_reader/reader_type.dart';
 import 'package:provider/provider.dart';
 
 class ImeiScanner extends StatefulWidget {
-  final Function(List<String>? scannedList)? onProceed;
+  final Function(List<String>? scannedList, {CameraDataModel? imageRawData})? onProceed;
   final VoidCallback? onTimeOut;
+  final ParserConfig config;
 
-  const ImeiScanner({super.key, this.onProceed, this.onTimeOut});
+  const ImeiScanner({super.key, this.onProceed, this.onTimeOut, required this.config});
 
   @override
   State<ImeiScanner> createState() => _ImeiScannerState();
@@ -43,13 +41,14 @@ class _ImeiScannerState extends State<ImeiScanner> {
           create: (_) => PageParamProvider(),
           child: ImeiSerialReader(
             configurationModel: ImeiSerialReaderConfig(
-              readerType: ReaderType.imeiReader.value,
+              readerType: widget.config.readerType.value,
               retryButtonText: 'Retry',
               doneButtonText: 'Done',
             ),
-            onDoneCallback: (List<String>? scannedList, {CameraImage? imageRawData}) {
-              widget.onProceed?.call(scannedList);
+            onDoneCallback: (scannedList, {cameraDataModel}) {
+              widget.onProceed?.call(scannedList, imageRawData: cameraDataModel);
             },
+            parserConfig: widget.config,
           ),
         ),
       ),