+++ lib/src/utils/media_upload/providers/media_upload_service_init_provider.dart
@@ -1,7 +1,7 @@
 import 'package:core_widgets/core_widgets.dart';
 import 'package:flutter_trc/src/libraries/shared_preferences/app_preferences.dart';
 import 'package:flutter_trc/src/modules/login/resources/login_types.dart';
-import 'package:flutter_trc/src/utils/media_upload/models/image_upload_service_type_enum.dart';
+import 'package:flutter_trc/src/utils/media_upload/models/media_upload_service_type_enum.dart';
 
 abstract class MediaUploadServiceInitProvider extends CshChangeNotifier {
   late final BaseService mediaUploadService;
@@ -13,6 +13,6 @@ abstract class MediaUploadServiceInitProvider extends CshChangeNotifier {
   void _initMediaService() {
     var loginType = AppPreferences.app.getLoginType();
     var loginTypeEnum = LoginTypes.fromValue(loginType ?? "");
-    mediaUploadService = ImageUploadServiceType.fromLoginType(loginTypeEnum).service;
+    mediaUploadService = MediaUploadServiceType.fromLoginType(loginTypeEnum).service;
   }
 }