+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # TRC Releases
 
+# v4.1.0+71 (17-03-2025)
+
+### Feature:
+
+* `QC/TRC - Manaul Testing - Serial Number Validation for laptops`[(SCM-8584)](https://reglobe.atlassian.net/browse/SCM-8584)
+* `QC - Data Wipe - Serial Number/Imei Validation flow`[(SCM-9042)](https://reglobe.atlassian.net/browse/SCM-9042)
+
+
 # v4.0.4+70 (07-03-2025)
 
 ### Hot fix: